meta_transfer_project/requirements.txt
torch>=2.0
torchvision
numpy
Pillow
tqdm
matplotlib
scipy
